Move ALL_PRODUCTS_PERMISSIONS back to models
